+++ packages/core/src/cli/validate/build-info-file.test.ts
@@ -95,6 +95,200 @@ const BUILD_INFO_NO_LAYOUT = {
   },
 };
 
+const BUILD_INFO_INDIVIDUAL_NO_LAYOUT = {
+  solcVersion: '0.8.9',
+  input: {
+    language: 'Solidity',
+    sources: {
+      'mypath/MyContract.sol': {
+        content: 'contract MyContract {}',
+      },
+      'mypath/MyContractV2.sol': {
+        content: 'contract MyContractV2 {}',
+      },
+    },
+    settings: {
+      outputSelection: {
+        'mypath/MyContract.sol': {
+          '': ['ast'],
+          '*': ['abi', 'evm.bytecode', 'evm.deployedBytecode', 'evm.methodIdentifiers', 'metadata'],
+        },
+        'mypath/MyContractV2.sol': {
+          '': ['ast'],
+          '*': ['abi', 'evm.bytecode', 'evm.deployedBytecode', 'evm.methodIdentifiers', 'metadata'],
+        },
+      },
+    },
+  },
+  output: {
+    sources: {
+      'mypath/MyContract.sol': {
+        ast: {},
+        id: 123,
+      },
+      'mypath/MyContractV2.sol': {
+        ast: {},
+        id: 456,
+      },
+    },
+  },
+};
+
+const BUILD_INFO_INDIVIDUAL_HAS_LAYOUT = {
+  solcVersion: '0.8.9',
+  input: {
+    language: 'Solidity',
+    sources: {
+      'mypath/MyContract.sol': {
+        content: 'contract MyContract {}',
+      },
+      'mypath/MyContractV2.sol': {
+        content: 'contract MyContractV2 {}',
+      },
+    },
+    settings: {
+      outputSelection: {
+        'mypath/MyContract.sol': {
+          '': ['ast'],
+          '*': ['abi', 'evm.bytecode', 'evm.deployedBytecode', 'evm.methodIdentifiers', 'metadata', 'storageLayout'],
+        },
+        'mypath/MyContractV2.sol': {
+          '': ['ast'],
+          '*': ['abi', 'evm.bytecode', 'evm.deployedBytecode', 'evm.methodIdentifiers', 'metadata', 'storageLayout'],
+        },
+      },
+    },
+  },
+  output: {
+    sources: {
+      'mypath/MyContract.sol': {
+        ast: {},
+        id: 123,
+      },
+      'mypath/MyContractV2.sol': {
+        ast: {},
+        id: 456,
+      },
+    },
+  },
+};
+
+const BUILD_INFO_PARTIAL_LAYOUT = {
+  solcVersion: '0.8.9',
+  input: {
+    language: 'Solidity',
+    sources: {
+      'mypath/MyContract.sol': {
+        content: 'contract MyContract {}',
+      },
+      'mypath/MyContractV2.sol': {
+        content: 'contract MyContractV2 {}',
+      },
+    },
+    settings: {
+      outputSelection: {
+        'mypath/MyContract.sol': {
+          '': ['ast'],
+          '*': ['abi', 'evm.bytecode', 'evm.deployedBytecode', 'evm.methodIdentifiers', 'metadata', 'storageLayout'],
+        },
+        'mypath/MyContractV2.sol': {
+          '': ['ast'],
+          '*': ['abi', 'evm.bytecode', 'evm.deployedBytecode', 'evm.methodIdentifiers', 'metadata'],
+        },
+      },
+    },
+  },
+  output: {
+    sources: {
+      'mypath/MyContract.sol': {
+        ast: {},
+        id: 123,
+      },
+      'mypath/MyContractV2.sol': {
+        ast: {},
+        id: 456,
+      },
+    },
+  },
+};
+
+const BUILD_INFO_PARTIAL_COMPILE = {
+  solcVersion: '0.8.9',
+  input: {
+    language: 'Solidity',
+    sources: {
+      'mypath/MyContract.sol': {
+        content: 'contract MyContract {}',
+      },
+      'mypath/MyContractV2.sol': {
+        content: 'contract MyContractV2 {}',
+      },
+    },
+    settings: {
+      outputSelection: {
+        'mypath/MyContract.sol': {
+          '': ['ast'],
+          '*': ['abi', 'evm.bytecode', 'evm.deployedBytecode', 'evm.methodIdentifiers', 'metadata', 'storageLayout'],
+        },
+        'mypath/MyContractV2.sol': {
+          '': [],
+          '*': [],
+        },
+      },
+    },
+  },
+  output: {
+    sources: {
+      'mypath/MyContract.sol': {
+        ast: {},
+        id: 123,
+      },
+      'mypath/MyContractV2.sol': {
+        ast: {},
+        id: 456,
+      },
+    },
+  },
+};
+
+const BUILD_INFO_NO_OUTPUT_SELECTION = {
+  solcVersion: '0.8.9',
+  input: {
+    language: 'Solidity',
+    sources: {
+      'mypath/MyContract.sol': {
+        content: 'contract MyContract {}',
+      },
+      'mypath/MyContractV2.sol': {
+        content: 'contract MyContractV2 {}',
+      },
+    },
+    settings: {
+      outputSelection: {
+        'mypath/MyContract.sol': {
+          '*': [],
+        },
+        'mypath/MyContractV2.sol': {
+          '': [],
+          '*': [],
+        },
+      },
+    },
+  },
+  output: {
+    sources: {
+      'mypath/MyContract.sol': {
+        ast: {},
+        id: 123,
+      },
+      'mypath/MyContractV2.sol': {
+        ast: {},
+        id: 456,
+      },
+    },
+  },
+};
+
 test.serial('get build info files - default hardhat', async t => {
   await fs.mkdir('artifacts/build-info', { recursive: true });
   await fs.mkdir('out/build-info', { recursive: true }); // should be ignored since it's empty
@@ -177,19 +371,72 @@ test.serial('dir does not exist', async t => {
 });
 
 test.serial('no build info files', async t => {
-  await fs.mkdir('empty-dir', { recursive: true });
-  await fs.writeFile('empty-dir/notjson.txt', 'abc');
+  const dir = 'empty-dir';
 
-  const error = await t.throwsAsync(getBuildInfoFiles('empty-dir'));
+  await fs.mkdir(dir, { recursive: true });
+  await fs.writeFile(`${dir}/notjson.txt`, 'abc');
+
+  const error = await t.throwsAsync(getBuildInfoFiles(dir));
   t.true(error?.message.includes('does not contain any build info files'));
 });
 
 test.serial('no storage layout', async t => {
-  await fs.mkdir('no-storage-layout', { recursive: true });
-  await fs.writeFile('no-storage-layout/build-info.json', JSON.stringify(BUILD_INFO_NO_LAYOUT));
+  const dir = 'no-storage-layout';
 
-  const error = await t.throwsAsync(getBuildInfoFiles('no-storage-layout'));
+  await fs.mkdir(dir, { recursive: true });
+  await fs.writeFile(`${dir}/build-info.json`, JSON.stringify(BUILD_INFO_NO_LAYOUT));
+
+  const error = await t.throwsAsync(getBuildInfoFiles(dir));
   t.true(error?.message.includes('does not contain storage layout'));
+});
+
+test.serial('individual output selections - no layout', async t => {
+  const dir = 'individual-no-layout';
+
+  await fs.mkdir(dir, { recursive: true });
+  await fs.writeFile(`${dir}/build-info.json`, JSON.stringify(BUILD_INFO_INDIVIDUAL_NO_LAYOUT));
+
+  const error = await t.throwsAsync(getBuildInfoFiles(dir));
+  t.true(error?.message.includes('does not contain storage layout'));
+});
+
+test.serial('individual output selections - has layout', async t => {
+  const dir = 'individual-has-layout';
+
+  await fs.mkdir(dir, { recursive: true });
+  await fs.writeFile(`${dir}/build-info.json`, JSON.stringify(BUILD_INFO_INDIVIDUAL_HAS_LAYOUT));
+
+  t.assert((await getBuildInfoFiles(dir)).length === 1);
+});
+
+test.serial('individual output selections - partial layout', async t => {
+  const dir = 'partial-layout';
+
+  await fs.mkdir(dir, { recursive: true });
+  await fs.writeFile(`${dir}/build-info.json`, JSON.stringify(BUILD_INFO_PARTIAL_LAYOUT));
+
+  const error = await t.throwsAsync(getBuildInfoFiles(dir));
+  t.true(error?.message.includes('does not contain storage layout'));
+});
+
+test.serial('individual output selections - partial compile', async t => {
+  const dir = 'partial-compile';
+
+  await fs.mkdir(dir, { recursive: true });
+  await fs.writeFile(`${dir}/build-info.json`, JSON.stringify(BUILD_INFO_PARTIAL_COMPILE));
+
+  const error = await t.throwsAsync(getBuildInfoFiles(dir));
+  t.true(error?.message.includes('is not from a full compilation'));
+});
+
+test.serial('no output selection', async t => {
+  const dir = 'no-output';
+
+  await fs.mkdir(dir, { recursive: true });
+  await fs.writeFile(`${dir}/build-info.json`, JSON.stringify(BUILD_INFO_NO_OUTPUT_SELECTION));
+
+  const error = await t.throwsAsync(getBuildInfoFiles(dir));
+  t.true(error?.message.includes('is not from a full compilation'));
 });
 
 function assertBuildInfoFiles(t: ExecutionContext, buildInfoFiles: BuildInfoFile[]) {